shellEntry: Update shell entry if browsers are disabled
Update the shell's search-entry hint text if browsers are
enabled or disabled via parental controls. The logic for
browsers being disabled or enabled, is driven by managing
"x-scheme-handler/http" content type in malcontent and
hence, is a deciding factor in these changes to select
relevant hint text for the search entry.

https://phabricator.endlessm.com/T26991