js: Use StPasswordEntry for password entry fields
Use the new StPasswordEntry for password entry fields
and remove all direct handling of clutter text of the entry
in shellEntry via clutter_text_set_password_char to show/hide
the password text. StPasswordEntry will provides a
peek-password-icon which will allow to show/hide the password
present in the field to the user in subsequent commits.

https://gitlab.gnome.org/GNOME/gnome-shell/merge_requests/619